OpenAI_SDK_Agents/6_model_configurations/pyproject.toml
[project]
name = "6-model-configurations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.2.3",
]
